requirements.txt
flask==2.1.0
flask-cors==3.0.10
opencv-python-headless==4.5.5.64
mysql-connector-python==8.0.28
numpy==1.22.3
python-dotenv==0.20.0
gunicorn==20.1.0  # For production deployment